+++ monitorcontrol/__init__.py
@@ -1,5 +1,5 @@
 from . import vcp  # noqa: F401
-from .vcp import VCPError, VCPIOError, VCPPermissionError  # noqa: F401
+from .vcp import vcp_codes, VCPError, VCPIOError, VCPPermissionError  # noqa: F401
 from .monitorcontrol import (  # noqa: F401
     get_monitors,
     Monitor,